Revert map sections
